# MRG, ENH: Speed up info copy