--- a/mne/io/fieldtrip/utils.py
+++ b/mne/io/fieldtrip/utils.py
@@ -309,9 +309,10 @@ def _process_channel_meg(cur_ch, grad):
     original_orientation = np.squeeze(grad['chanori'][chan_idx_in_grad, :])
     try:
         orientation = rotation3d_align_z_axis(original_orientation).T
-        orientation = orientation.flatten()
     except AssertionError:
-        orientation = np.eye(4, 4).flatten()
+        orientation = np.eye(3)
+    assert orientation.shape == (3, 3)
+    orientation = orientation.flatten()
     chanunit = grad['chanunit'][chan_idx_in_grad]
 
     cur_ch['loc'] = np.hstack((position, orientation))