@agramfort 2020-03-19
thx @larsoner